Bugfix: Change 'after suite' hooks to 'after all'
The before/after suite hooks don't allow you to specify a filter and therefore
they execute on every spec run.  This was especially a problem for the Stripe
plugin which was complaining about not having an API key, even though the spec
we were running wasn't even using Stripe.

Changing to a 'before all' hook works, but has the slight downside of the fact
that it may be run multiple times if there are multiple files with `stripe:
true` since that hook is run before/after every file.

--------------------------------------------------------------------------------
Change-Id:     I98216754b009d658b7d4aca80a1e08c39a1a3d5c
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>