Feature: Add all Rails Engine app paths to the LOAD_PATH
When working with a Rails engine, due to the stupid autoloading crap, we can't
explicitly monkey-patch anything in the app directory.

The workaround for this is to explicitly `require` the file we want to
monkeypatch before we actually perform the monkeypatch.  Unfortunately doing
this requires weird hacky-looking code such as:

    require Rails.root.join('..', '..', 'app', 'controllers', 'engine_name', etc)

So what this feature does is add all the subdirectories under the engine's `app`
directory to the `LOAD_PATH` so that we can do something like this instead:

    require 'engine_name/controller_name'

--------------------------------------------------------------------------------
Change-Id:     I883e69a3aba2e1fd78c29c3ee9101d06f78cb8c0
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>