Feature: Set verify_doubled_constant_names
By default we want to make sure that if we attempt to verify a double against
a class which does not exist it should raise an error.  In other words:

```ruby
class MyClass
  def do_stuff
    my_method
  end

  def my_method
  end
end

describe MyClass do
  it 'verifies the double' do
    my_double = instance_double(MyClass)

    my_double.do_stuff

    expect(my_double).to receive(:my_method)
  end
end
```

The above will pass.

```ruby
describe MyClass do
  it 'verifies the double' do
    my_double = instance_double(MyClass)

    my_double.do_stuff

    expect(my_double).to receive(:my_method)
  end
end
```

The above will fail when creating the instance double not when calling
'do_stuff', because 'MyClass' is not loaded.

--------------------------------------------------------------------------------
Change-Id:     Ie5a67ac2665ef11dc80ed5f1d598680a727a49e9
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>