Feature: Add VCR matcher which ignores trailing ID from URI
This adds two matchers that can be used:

  * uri_without_trailing_id
  * uri_without_trailing_guid

The `uri_without_trailing_id` will match URIs that end in all digits.

The `uri_without_trailing_guid` will match URIs which end in exactly 32
hexadecimal numbers.

If the URIs do not match, then the entire URI will be compared against that in
the VCR cassette, otherwise, everything except the ID portion will be compared.

--------------------------------------------------------------------------------
Change-Id:     If466f7d030cdf92fbdd407222f64f2cacba090d5
Signed-off-by: Jeff Felchner and Drew Montgomery <pair+hhd418-jfelchner@example.com>